Renamed package to me.executed.scoreboardplus
